Another big battle UI overhaul.
Definitely of the opinion that putting much thought into the UI design is a waste of time as it is so far from my skillset. However, there's still quite a lot of essential work to do, some of which this commit covers - the display of calm and strife resource bars, including Jack's unique case.